New commands update
I have updated the command to get the day of the month to use systemTime instead of ExileServerStartTime. This means I can drop the overrides for GADDvent to work. It also means that people in different timezones can access their daily gifts at 1 min past midnight their time, for that day if they so desire.